// UI-MySql/src/app/models/employee.ts
export interface Employee {
    Id: any;
    Title: string;
    EmployeeName: string;
    EmployeeStatusId: number;
    ReportingTo: number;
    GenderId: number;
    MaritalStatusId: number;
    BranchId: number;
    OfficeEmail: string;
    OfficeMobileNumber: string;
    EmployeeId: string;
    PersonalMobileNumber: string;
    PersonalEmail: string;
    DateOfBirth: string;
    DesignationId: number;
    DepartmentId: number;
    FatherName: string;
    MotherName: string;
    EmployementTypeId: number;
    Aadhar: string;
    PAN: string;
    ReligionId: number;
    BloodGroupId: number;
    EmergencyContact: string;
    UAN: string;
    ESI: string;
    DateOfJoining: String;
    CheckInTime: string;
    CheckOutTime: string;
    DateOfExit?: String;
    AccountName: string;
    PresentAddressLine1: string;
    PresentAddressLine2?: string;
    PresentAddressCity: string;
    PresentAddressPincode: string;
    PresentAddressStateId: number;
    PresentAddressCountryId: number;
    IsPermanentSameAsPresent: boolean;
    PermanentAddressLine1: string;
    PermanentAddressLine2: string;
    PermanentAddressCity: string;
    PermanentAddressPincode: string;
    PermanentAddressStateId: number;
    PermanentAddressCountryId: number;
    BankAccountNumber: string;
    BankAccountName: string;
    BankId: number;
    BankIfscCode: string;
    BankAccountTypeId: number;
    BankBranchName: string;
}